default swap logic. pay lease mock. liquidity ready to estimate payment

+++ src/DefaultStrategy.sol
@@ -2,13 +2,13 @@
 pragma solidity ^0.8.24;
 
 interface IStrategy {
-    function getFee() external view returns (uint256);
+    function getFee() external view returns (uint128);
 }
 
 contract DefaultStrategy is IStrategy {
-    uint256 private FEE = 3000;
+    uint128 private FEE = 3000;
 
-    function getFee() external view returns (uint256) {
+    function getFee() external view returns (uint128) {
         return FEE;
     }
 }
